Add 'concat' and 'reverse' methods to array type; closes #4

+++ tests/suites/test_objects.c
@@ -13,7 +13,7 @@ void
 test_object_creation ()
 {
   proto_object_t *object;
-  int value = 10;
+  short int value = 10;
 
   describe ("Create and destroy object");
   object = proto_init_object ();
@@ -21,7 +21,7 @@ test_object_creation ()
   object->set_own_property (object, "testing", &value);
   should_be_true (object->has_own_property (object, "testing"));
   should_be_false (object->has_own_property (object, "another_key"));
-  should_equal (*((int *) object->get_own_property (object, "testing")), value);
+  should_equal (*((short int *) object->get_own_property (object, "testing")), value);
   should_equal (object->get_own_property (object, "another_key"), NULL);
   proto_del_object (object);
 }
@@ -30,7 +30,7 @@ void
 test_object_reassignment_of_key ()
 {
   proto_object_t *object;
-  int value_a = 10, value_b = 20;
+  short int value_a = 10, value_b = 20;
 
   describe ("Create object and reassign a value to the same key");
   object = proto_init_object ();
@@ -38,10 +38,10 @@ test_object_reassignment_of_key ()
   object->set_own_property (object, "testing", &value_a);
   should_be_true (object->has_own_property (object, "testing"));
   should_be_false (object->has_own_property (object, "another_key"));
-  should_equal (*((int *) object->get_own_property (object, "testing")), value_a);
+  should_equal (*((short int *) object->get_own_property (object, "testing")), value_a);
   should_equal (object->get_own_property (object, "another_key"), NULL);
   object->set_own_property (object, "testing", &value_b);
-  should_equal (*((int *) object->get_own_property (object, "testing")), value_b);
+  should_equal (*((short int *) object->get_own_property (object, "testing")), value_b);
   proto_del_object (object);
 }
 
@@ -49,18 +49,18 @@ void
 test_object_with_pointer ()
 {
   proto_object_t *object;
-  int *pointer, *deleted;
+  short int *pointer, *deleted;
 
   describe ("Create object and assign a key to a pointer");
   object = proto_init_object ();
   should_be_true (object != NULL);
-  pointer = (int *) malloc (sizeof (int));
+  pointer = (short int *) malloc (sizeof (short int));
   *pointer = 200;
   object->set_own_property (object, "pointer", pointer);
   should_be_true (object->has_own_property (object, "pointer"));
-  should_equal ((int *) object->get_own_property (object, "pointer"), pointer);
-  should_equal (*(int *) object->get_own_property (object, "pointer"), *pointer);
-  deleted = (int *) object->del_own_property (object, "pointer");
+  should_equal ((short int *) object->get_own_property (object, "pointer"), pointer);
+  should_equal (*(short int *) object->get_own_property (object, "pointer"), *pointer);
+  deleted = (short int *) object->del_own_property (object, "pointer");
   free (deleted);
   proto_del_object (object);
 }
@@ -69,20 +69,20 @@ void
 test_object_with_multiple_keys ()
 {
   proto_object_t *object;
-  int value_a = 10, value_b = 20, value_c = 30;
+  short int value_a = 10, value_b = 20, value_c = 30;
 
   describe ("Create object and assign multiple keys");
   object = proto_init_object ();
   should_be_true (object != NULL);
   object->set_own_property (object, "a", &value_a);
   should_be_true (object->has_own_property (object, "a"));
-  should_equal (*((int *) object->get_own_property (object, "a")), value_a);
+  should_equal (*((short int *) object->get_own_property (object, "a")), value_a);
   object->set_own_property (object, "b", &value_b);
   should_be_true (object->has_own_property (object, "b"));
-  should_equal (*((int *) object->get_own_property (object, "b")), value_b);
+  should_equal (*((short int *) object->get_own_property (object, "b")), value_b);
   object->set_own_property (object, "c", &value_c);
   should_be_true (object->has_own_property (object, "c"));
-  should_equal (*((int *) object->get_own_property (object, "c")), value_c);
+  should_equal (*((short int *) object->get_own_property (object, "c")), value_c);
   proto_del_object (object);
 }
 
@@ -90,23 +90,23 @@ void
 test_object_with_colliding_keys ()
 {
   proto_object_t *object;
-  int value_a = 10, value_b = 20, value_c = 30, value_d = 40;
+  short int value_a = 10, value_b = 20, value_c = 30, value_d = 40;
 
   describe ("Create object and assign colliding keys");
   object = proto_init_object ();
   should_be_true (object != NULL);
   object->set_own_property (object, "computer", &value_a);
   should_be_true (object->has_own_property (object, "computer"));
-  should_equal (*((int *) object->get_own_property (object, "computer")), value_a);
+  should_equal (*((short int *) object->get_own_property (object, "computer")), value_a);
   object->set_own_property (object, "programming", &value_b);
   should_be_true (object->has_own_property (object, "programming"));
-  should_equal (*((int *) object->get_own_property (object, "programming")), value_b);
+  should_equal (*((short int *) object->get_own_property (object, "programming")), value_b);
   object->set_own_property (object, "testing", &value_c);
   should_be_true (object->has_own_property (object, "testing"));
-  should_equal (*((int *) object->get_own_property (object, "testing")), value_c);
+  should_equal (*((short int *) object->get_own_property (object, "testing")), value_c);
   object->set_own_property (object, "go", &value_d);
   should_be_true (object->has_own_property (object, "go"));
-  should_equal (*((int *) object->get_own_property (object, "go")), value_d);
+  should_equal (*((short int *) object->get_own_property (object, "go")), value_d);
   proto_del_object (object);
 }
 
@@ -114,28 +114,28 @@ void
 test_object_deletion_of_key ()
 {
   proto_object_t *object;
-  int value_a = 10, value_b = 20, value_c = 30, value_d = 40, deleted;
+  short int value_a = 10, value_b = 20, value_c = 30, value_d = 40, deleted;
 
   describe ("Create object, assign keys and delete them");
   object = proto_init_object ();
   should_be_true (object != NULL);
   object->set_own_property (object, "computer", &value_a);
   should_be_true (object->has_own_property (object, "computer"));
-  should_equal (*((int *) object->get_own_property (object, "computer")), value_a);
+  should_equal (*((short int *) object->get_own_property (object, "computer")), value_a);
   object->set_own_property (object, "programming", &value_b);
   should_be_true (object->has_own_property (object, "programming"));
-  should_equal (*((int *) object->get_own_property (object, "programming")), value_b);
+  should_equal (*((short int *) object->get_own_property (object, "programming")), value_b);
   object->set_own_property (object, "testing", &value_c);
   should_be_true (object->has_own_property (object, "testing"));
-  should_equal (*((int *) object->get_own_property (object, "testing")), value_c);
+  should_equal (*((short int *) object->get_own_property (object, "testing")), value_c);
   object->set_own_property (object, "go", &value_d);
   should_be_true (object->has_own_property (object, "go"));
-  should_equal (*((int *) object->get_own_property (object, "go")), value_d);
+  should_equal (*((short int *) object->get_own_property (object, "go")), value_d);
 
   // Delete keys
 
   // Delete "computer"
-  deleted = *(int *) object->del_own_property (object, "computer");
+  deleted = *(short int *) object->del_own_property (object, "computer");
   should_equal (deleted, value_a);
   should_be_false (object->has_own_property (object, "computer"));
   should_be_true (object->has_own_property (object, "programming"));
@@ -143,20 +143,20 @@ test_object_deletion_of_key ()
   should_be_true (object->has_own_property (object, "go"));
 
   // Delete "programming"
-  deleted = *(int *) object->del_own_property (object, "programming");
+  deleted = *(short int *) object->del_own_property (object, "programming");
   should_equal (deleted, value_b);
   should_be_false (object->has_own_property (object, "programming"));
   should_be_true (object->has_own_property (object, "testing"));
   should_be_true (object->has_own_property (object, "go"));
 
   // Delete "testing"
-  deleted = *(int *) object->del_own_property (object, "testing");
+  deleted = *(short int *) object->del_own_property (object, "testing");
   should_equal (deleted, value_c);
   should_be_false (object->has_own_property (object, "testing"));
   should_be_true (object->has_own_property (object, "go"));
 
   // Delete "go"
-  deleted = *(int *) object->del_own_property (object, "go");
+  deleted = *(short int *) object->del_own_property (object, "go");
   should_equal (deleted, value_d);
   should_be_false (object->has_own_property (object, "go"));
 
@@ -167,7 +167,7 @@ void
 test_object_chain_calls ()
 {
   proto_object_t *object_a, *object_b, *object_c;
-  int value = 10;
+  short int value = 10;
 
   describe ("Create three objects, assign keys for each object, in a chain, and then call it");
   object_a = proto_init_object ();
@@ -179,12 +179,12 @@ test_object_chain_calls ()
   should_be_true (object_a->has_own_property (object_a, "b"));
   should_be_true (object_b->has_own_property (object_b, "c"));
   should_be_true (object_c->has_own_property (object_c, "value"));
-  should_equal (*(int *) object_a->chain (object_a, "b.c.value"), value);
+  should_equal (*(short int *) object_a->chain (object_a, "b.c.value"), value);
   should_equal (object_a->chain (object_a, "b.c"), (void *) object_c);
   should_equal (object_a->chain (object_a, "b"), (void *) object_b);
-  should_equal (*(int *) object_b->chain (object_b, "c.value"), value);
+  should_equal (*(short int *) object_b->chain (object_b, "c.value"), value);
   should_equal (object_b->chain (object_b, "c"), (void *) object_c);
-  should_equal (*(int *) object_c->chain (object_c, "value"), value);
+  should_equal (*(short int *) object_c->chain (object_c, "value"), value);
   proto_del_object (object_a);
   proto_del_object (object_b);
   proto_del_object (object_c);
@@ -193,8 +193,8 @@ test_object_chain_calls ()
 static void *
 mock_function (void *arguments)
 {
-  int value = *(int *) arguments;
-  int *sum = malloc (sizeof (int));
+  short int value = *(short int *) arguments;
+  short int *sum = malloc (sizeof (short int));
   *sum = value + 10;
   return sum;
 }
@@ -203,13 +203,13 @@ void
 test_object_execute_function ()
 {
   proto_object_t *object;
-  int value = 10, *sum;
+  short int value = 10, *sum;
 
   describe ("Create objects, assign a function to a key and then execute it");
   object = proto_init_object ();
   object->set_own_property (object, "function", &mock_function);
   should_equal (object->get_own_property (object, "function"), &mock_function);
-  sum = (int *) object->execute_property (object, "function", &value);
+  sum = (short int *) object->execute_property (object, "function", &value);
   should_equal (*sum, value + 10);
   free (sum);
   proto_del_object (object);
